fix:AI API수정
